Move hex_codes adn messages to own file
