+++ src/hex_codes.h
@@ -1,0 +1,7 @@
+enum HexCodes{
+  ACK = 0x06,
+  NACK = 0x15,
+  STX = 0x02,
+  ETX = 0x03,
+  PIPE = 0x7C
+};
