+++ src/transbank.h
@@ -6,23 +6,15 @@
 #include <stdlib.h>
 #include <string.h>
 #include <libserialport.h>
+#include "message.h"
 
-enum tbk_return{
+enum TbkReturn{
     TBK_OK = 0,
     TBK_NOK = -1
 };
 
-typedef struct message_t Message;
-
-struct message_t {
-  int payloadSize;
-  int responseSize;
-  int retries;
-  char* payload;
-};
-
-extern enum tbk_return open_port(char* portName, int baudrate);
-extern enum tbk_return polling();
+extern enum TbkReturn open_port(char* portName, int baudrate);
+extern enum TbkReturn polling();
 extern int set_normal_mode();
 extern int close_port();
