+++ src/transbank_serial_utils.h
@@ -3,14 +3,6 @@
 
 #include "transbank.h"
 
-enum HexCodes{
-  ACK = 0x06,
-  NACK = 0x15,
-  STX = 0x02,
-  ETX = 0x03,
-  PIPE = 0x7C
-};
-
 extern char* list_ports();
 extern char* get_port_name(struct sp_port *port);
 extern int read_bytes(struct sp_port *port, char* buf, Message message);
